refactor: AuthErrorCode package 이동
